Use go 1.23 iterators.
Add All() to Sequence to return all the positions and digits.
Add Backward() to FiniteSequence to return the same from end to beginning.
Add All() to Positions to return all the ranges.
Add Matches() to return all 0 based positions where pattern is found in s.
Add BackwardMatches() to return the same from last to first.